Refactor logic to parse ScanCode results to a separate file
In order to integrate ClearlyDefined, such result files will be
downloaded from a server. Therefore, it makes sense to separate the
logic to parse results from the logic to invoke ScanCode.

Add a function to generate a ScannerDetails object from a result, too.
